GLM refactoring in progress, single test passes
