+++ tests/test_simple_inference.py
@@ -4,7 +4,9 @@ from collections import Counter
 
 import numpy as np
 
-from alasmc.main import ModelSelectionSMC, ModelKernel, get_model_id, binomial_logit_logll, normal_prior
+from alasmc.main import ModelSelectionSMC, ModelKernel, normal_prior
+from alasmc.GLM import LogisticGLM
+from alasmc.utilities import get_model_id
 
 def create_data(n_covariates, n_active):
     n_covariates = 5
@@ -31,14 +33,7 @@ def easy_data():
     model_init = np.array([False] * n_covariates)
     model_init[np.random.choice(n_covariates)] = True
 
-    def db(linpred):
-        return 1 / (1 + np.exp(-linpred))
-
-    def d2b(linpred):
-        p = 1 / (1 + np.exp(-linpred))
-        return p * (1 - p)
-
-    smc = ModelSelectionSMC(X, y, likelihood=binomial_logit_logll, db=db, d2b=d2b,
+    smc = ModelSelectionSMC(X, y, glm=LogisticGLM,
                             coef_init=np.array([0] * n_covariates), model_init=model_init, coef_prior=normal_prior,
                             kernel=kernel, kernel_steps=5, particle_number=particle_number, verbose=True)
     return smc, beta_true
